renamed components to convention
fixed name mismatches
