+++ react-app/src/App.js
@@ -10,7 +10,7 @@ import EditSongForm from "./components/SongFolders/EditSongForm";
 import { getAllSongs } from "./store/song";
 import { getAllPlaylists } from "./store/playlist";
 import SingleSong from "./components/SongFolders/SingleSong";
-import Audio from "./components/AudioPlayer";
+import Audio from "./components/AudioPlayer/Audio";
 import SplashPage from "./components/SplashPage";
 import PlaylistsPage from "./components/PlaylistFolders/PlaylistsPage";
 import LibraryPage from "./components/LibraryPage";
